Move dynamic table functions to participant filter.
