+++ lib/amd/src/filter.js
@@ -22,7 +22,6 @@
  */
 
 import CourseFilter from 'core/local/filter/filtertypes/courseid';
-import * as DynamicTable from 'core_table/dynamic';
 import GenericFilter from 'core/local/filter/base';
 import {get_strings as getStrings} from 'core/str';
 import Notification from 'core/notification';
@@ -35,12 +34,10 @@ import jQuery from 'jquery';
 /**
  * Initialise the participants filter on the element with the given id.
  *
- * @param {String} filterRegionId The id for the filter element.
+ * @param {HTMLElement} filterSet The filter element.
  * @param {Function} applyCallback Callback function when updateTableFromFilter
  */
-export const init = (filterRegionId, applyCallback) => {
-    // Keep a reference to the filterset.
-    const filterSet = document.querySelector(`#${filterRegionId}`);
+export const init = (filterSet, applyCallback) => {
 
     // Keep a reference to all of the active filters.
     const activeFilters = {
@@ -435,18 +432,6 @@ export const init = (filterRegionId, applyCallback) => {
         return legendStrings;
     };
 
-    const setDynamicTable = () => {
-        const tableRoot = DynamicTable.getTableFromId(filterSet.dataset.tableRegion);
-        const initialFilters = DynamicTable.getFilters(tableRoot);
-        if (initialFilters) {
-            const initialFilterPromise = new Pending('core/filter:setFilterFromConfig');
-            // Apply the initial filter configuration.
-            setFilterFromConfig(initialFilters)
-                .then(() => initialFilterPromise.resolve())
-                .catch();
-        }
-    };
-
     // Add listeners for the main actions.
     filterSet.querySelector(Selectors.filterset.region).addEventListener('click', e => {
         if (e.target.closest(Selectors.filterset.actions.addRow)) {
@@ -493,19 +478,4 @@ export const init = (filterRegionId, applyCallback) => {
         filterSet.dataset.filterverb = e.target.value;
     });
 
-    // Set up by result type.
-    // let useDynamicTable = false;
-    // switch (resultType) {
-    //     case 'DynamicTable':
-    //         setDynamicTable();
-    //         useDynamicTable = true;
-    //         break;
-    //     case 'QbankTable': {
-    //         const filtertypeselect = filterSet.querySelector(Selectors.filter.fields.type);
-    //         filtertypeselect.value = 'category';
-    //         const filter = filterSet.querySelector(Selectors.filter.region);
-    //         addFilter(filter, filtertypeselect.value);
-    //         break;
-    //     }
-    // }
 };
